🔨 refactor: LetterWrite 페이지 리팩토링
